block info, action handling
