+++ chain/chain/src/store_validator/validate.rs
@@ -1,8 +1,8 @@
 use crate::StoreValidator;
 
 use near_primitives::block::{Block, BlockHeader, Tip};
+use near_primitives::epoch_block_info::BlockInfo;
 use near_primitives::epoch_info::EpochInfo;
-use near_primitives::epoch_manager::block_info::BlockInfo;
 use near_primitives::hash::CryptoHash;
 use near_primitives::shard_layout::{get_block_shard_uid, ShardUId};
 use near_primitives::sharding::{ChunkHash, ShardChunk, StateSyncInfo};
