+++ core/primitives/src/lib.rs
@@ -14,6 +14,7 @@ pub mod block_body;
 pub mod block_header;
 pub mod challenge;
 pub mod congestion_info;
+pub mod epoch_block_info;
 pub mod epoch_info;
 pub mod epoch_manager;
 pub mod epoch_sync;
