+++ core/schema-checker/schema-checker-macro/src/lib.rs
@@ -16,7 +16,6 @@ mod helper {
         let input = parse_macro_input!(input as DeriveInput);
         let name = &input.ident;
         let info_name = quote::format_ident!("{}_INFO", name);
-
         let type_id = quote! { std::any::TypeId::of::<#name>() };
         let info = match &input.data {
             Data::Struct(data_struct) => {
