+++ core/store/Cargo.toml
@@ -103,3 +103,9 @@ nightly = [
   "nightly_protocol",
 ]
 statelessnet_protocol = ["near-primitives/statelessnet_protocol"]
+protocol_schema = [
+  "near-crypto/protocol_schema",
+  "near-primitives/protocol_schema",
+  "near-parameters/protocol_schema",
+  "near-schema-checker-lib/protocol_schema",
+]
