+++ tools/protocol-schema-check/res/protocol_schema.toml
@@ -3,10 +3,11 @@ AccessKeyPermission = 885623561
 Account = 358811118
 AccountV2 = 337859929
 AccountVersion = 4249996519
+Action = 2541352025
 ActionCosts = 3115555891
 ActionError = 4217425219
 ActionErrorKind = 1632922469
-ActionReceipt = 2761595564
+ActionReceipt = 2414678051
 ActionsValidationError = 1053886215
 AddKeyAction = 356099649
 Approval = 593918844
@@ -30,23 +31,27 @@ BlockHeaderV1 = 3090297699
 BlockHeaderV2 = 765369625
 BlockHeaderV3 = 2075503242
 BlockHeaderV4 = 254520801
+BlockInfo = 80746680
+BlockInfoV1 = 735547821
+BlockInfoV2 = 1224525771
 BlockV1 = 3852808866
 BlockV2 = 665216644
 BlockV3 = 1423246645
 BlockV4 = 2723276051
+BlockWithChangesInfo = 887507517
 BufferedReceiptIndices = 897024511
-Challenge = 3534218232
-ChallengeBody = 3092587692
+Challenge = 3436852083
+ChallengeBody = 1162346392
 ChunkEndorsement = 1642962729
 ChunkEndorsementInner = 2425301775
 ChunkEndorsementV1 = 2471185877
 ChunkExtraV1 = 774877102
 ChunkHash = 1471814478
 ChunkHashHeight = 825215623
-ChunkProofs = 494709052
-ChunkState = 738787664
+ChunkProofs = 1298994151
+ChunkState = 2225323962
 ChunkStateTransition = 307448170
-ChunkStateWitness = 3391691684
+ChunkStateWitness = 1299024010
 ChunkStateWitnessAck = 177881908
 ChunkStats = 4176245277
 CompilationError = 738158707
@@ -59,9 +64,10 @@ CurrentEpochValidatorInfo = 1129345174
 DataReceipt = 2506806701
 DataReceiver = 1715762664
 DelayedReceiptIndices = 1315689119
-DelegateAction = 2467576519
+DelegateAction = 3898002655
 DeleteAccountAction = 3244670577
 DeleteKeyAction = 1374597333
+DeployContractAction = 2972267833
 Direction = 1296680832
 ED25519PublicKey = 213018126
 EncodedChunkStateWitness = 329848903
@@ -84,26 +90,34 @@ ExecutionOutcomeWithIdAndProof = 1300070947
 ExecutionOutcomeWithProof = 3663131288
 ExecutionStatus = 1810006625
 ExtCosts = 836392984
+FetchingStateStatus = 2204896805
+FlatStateChanges = 2811133731
+FlatStateDeltaMetadata = 3401366797
 FlatStateValue = 83834662
+FlatStorageCreationStatus = 3717607657
+FlatStorageReadyStatus = 677315221
+FlatStorageStatus = 1026335026
 FunctionCallAction = 2405840012
 FunctionCallError = 3652274053
 FunctionCallPermission = 1517509673
 HostError = 3173968216
+IgnoredVecU8 = 1855789801
 IntegerOverflowError = 2542362165
 InvalidAccessKeyError = 2954698659
 InvalidTxError = 1219344901
+KeyForFlatStateDelta = 2002998927
 LegacyAccount = 1291371319
-MaybeEncodedShardChunk = 3327571706
+MaybeEncodedShardChunk = 901175937
 MerklePathItem = 2615629611
 MessageDiscriminant = 3240833245
 MethodResolveError = 1206790835
 MissingTrieValueContext = 2666011379
 NextEpochValidatorInfo = 236248764
-NonDelegateAction = 3063046120
-PartialEncodedChunk = 1134428818
+NonDelegateAction = 3255205790
+PartialEncodedChunk = 1151962180
 PartialEncodedChunkPart = 194051090
-PartialEncodedChunkV1 = 3307106147
-PartialEncodedChunkV2 = 4189985881
+PartialEncodedChunkV1 = 1656475386
+PartialEncodedChunkV2 = 2918315046
 PartialEncodedStateWitness = 1465562178
 PartialEncodedStateWitnessInner = 3195106273
 PartialState = 3772957669
@@ -115,16 +129,18 @@ PromiseYieldTimeout = 3189361393
 PublicKey = 601042198
 RawStateChange = 1346612669
 RawStateChangesWithTrieKey = 881872573
-Receipt = 2240655747
-ReceiptEnum = 1757078716
-ReceiptProof = 4225949523
-ReceiptV0 = 2750910817
-ReceiptV1 = 152025219
+RawTrieNode = 4239211001
+RawTrieNodeWithSize = 1474149765
+Receipt = 2916802703
+ReceiptEnum = 3157292228
+ReceiptProof = 1019992812
+ReceiptV0 = 3604411866
+ReceiptV1 = 2994842769
 ReceiptValidationError = 551721215
 ReceivedData = 3601438283
 Secp256K1PublicKey = 4117078281
 Secp256K1Signature = 3687154735
-ShardChunk = 15190413
+ShardChunk = 834871798
 ShardChunkHeader = 4215449923
 ShardChunkHeaderInner = 3760333502
 ShardChunkHeaderInnerV1 = 680913470
@@ -133,28 +149,31 @@ ShardChunkHeaderInnerV3 = 3883121745
 ShardChunkHeaderV1 = 1805459603
 ShardChunkHeaderV2 = 3706194757
 ShardChunkHeaderV3 = 2763275079
-ShardChunkV1 = 177872429
-ShardChunkV2 = 3441083084
+ShardChunkV1 = 1814805625
+ShardChunkV2 = 1857597167
 ShardProof = 2773021473
 ShardUId = 2410086023
 Signature = 3997391707
-SignedDelegateAction = 2108906857
+SignedDelegateAction = 2482265228
 SignedTransaction = 3898692301
 SlashState = 3264273950
 SlashedValidator = 2601657743
 StakeAction = 2002027105
 StateChangeCause = 3313733177
-StateChangesForResharding = 881554139
+StateChangesForResharding = 3565001667
 StorageError = 1838871872
 StoredChunkStateTransitionData = 516372819
 String = 2587724713
 Tip = 305642482
-TransactionReceipt = 3690012777
-TransactionV0 = 3679872072
-TransactionV1 = 1276458085
+TransactionReceipt = 968816131
+TransactionV0 = 197396442
+TransactionV1 = 2594686420
 TransferAction = 1078380396
+TrieChanges = 3833039794
 TrieKey = 768968236
 TrieQueueIndices = 2601394796
+TrieRefcountAddition = 2117109883
+TrieRefcountSubtraction = 2150368599
 TxExecutionError = 706862037
 ValidatorKickoutReason = 2362237969
 ValidatorKickoutView = 2660746751
