+++ tools/speedy_sync/src/main.rs
@@ -8,8 +8,8 @@ use near_epoch_manager::types::EpochInfoAggregator;
 use near_epoch_manager::EpochManager;
 use near_primitives::block::Block;
 use near_primitives::block_header::BlockHeader;
+use near_primitives::epoch_block_info::BlockInfo;
 use near_primitives::epoch_info::EpochInfo;
-use near_primitives::epoch_manager::block_info::BlockInfo;
 use near_primitives::epoch_manager::AGGREGATOR_KEY;
 use near_primitives::hash::CryptoHash;
 use near_primitives::merkle::PartialMerkleTree;
